Добавление в список базовой валюты

+++ currency-converter-back/src/main/java/org/superngb/currencyconverterback/dto/CbCurrencyDtoModel.java
@@ -2,20 +2,28 @@ package org.superngb.currencyconverterback.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public record CbCurrencyDtoModel(
         @JsonProperty("ID")
         String id,
+
         @JsonProperty("NumCode")
         String numCode,
+
         @JsonProperty("CharCode")
         String charCode,
+
         @JsonProperty("Nominal")
         Integer nominal,
+
         @JsonProperty("Name")
         String name,
+
         @JsonProperty("Value")
-        Double value,
+        BigDecimal value,
+
         @JsonProperty("Previous")
-        Double previous
+        BigDecimal previous
 ) {
 }
